Backlight Mode Daily Sync
The backlight mode should synchronize with the remote control on a daily basis upon establishing a connection. This procedure is necessary to address potential clock drift within the remote control and its inability to retain backlight mode settings after entering a deep sleep state.

Test: manual
Bug: 339346781
Flag: EXEMPT config is not be enabled. It will be enabled on devices per device capabilities.
Change-Id: Id5e2885cea226f2cd6475e69c13d402329554be2